split script into three seperate js files to track whats happening more easily
