+++ README.md
@@ -1,0 +1,55 @@
+# q
+
+## [The Deployed Page](https://awesome-foursome.github.io/boardgame-bliss/)
+
+## The Project
+
+w
+
+## Link & Contents
+
+- [Installation](#installation)
+- [Usage](#usage)
+- [Future Development](#future-development)
+- [Credits](#credits)
+- [Contributing](#contributing)
+- [Questions and Feedback](#questions-and-feedback)
+- [Tests](#tests)
+- [License](#license)
+
+## Image
+
+![Picture of the Homepage](./assets/images/homepage-deployment-screenshot.png)
+
+## Installation
+
+e
+
+## Usage
+
+r
+
+## Future Development
+
+t
+
+## Credits
+
+y
+
+## Contributing
+
+u
+
+## Questions and Feedback
+
+i
+
+## Tests
+
+o
+
+## License
+
+Apache License
+
